Added lines to prevent unnecessary rpm switching
Overall test class now remembers previous test mode. If previous test is
also a capacitance test then the command to set the 4225-RPM switches is
ignored, as they will already be in the correct position. When IV test
is added this will allow the switches to be changed to SMU mode, again
only once when needed.